requirements.txt
streamlit
requests==2.32.3
pandas
scikit-learn==1.6.1
streamlit-folium
numpy
python-dotenv==1.0.1
transformers==4.48.1
huggingface-hub==0.27.1